+++ src/julia/Julia.cpp
@@ -50,7 +50,7 @@ int petri_copy(int const handle)
 {
     SANITY_HANDLE(handle, -1);
 
-    g_petri_nets.push_back(std::make_unique<PetriNet>(*g_petri_nets[handle]));
+    g_petri_nets.push_back(std::make_unique<PetriNet>(*g_petri_nets[size_t(handle)]));
     return int(g_petri_nets.size() - 1u);
 }
 
@@ -58,7 +58,7 @@ int petri_copy(int const handle)
 bool petri_reset(int const handle)
 {
     SANITY_HANDLE(handle, false);
-    g_petri_nets[handle]->reset();
+    g_petri_nets[size_t(handle)]->reset();
     return true;
 }
 
@@ -69,7 +69,7 @@ bool petri_is_empty(int const handle, bool* empty)
     if (empty == nullptr)
         return false;
 
-    *empty = g_petri_nets[handle]->isEmpty();
+    *empty = g_petri_nets[size_t(handle)]->isEmpty();
     return true;
 }
 
@@ -82,7 +82,7 @@ bool petri_editor(int const handle)
     SANITY_HANDLE(handle, false);
 
     Application application(800, 600, "Timed Petri Net Editor");
-    PetriEditor editor(application.renderer(), *g_petri_nets[handle]);
+    PetriEditor editor(application.renderer(), *g_petri_nets[size_t(handle)]);
     editor.bgColor = sf::Color(255,255,255,255);
 
     try
@@ -103,7 +103,7 @@ bool petri_editor(int const handle)
 int petri_count_places(int const handle)
 {
     SANITY_HANDLE(handle, -1);
-    return int(g_petri_nets[handle]->places().size());
+    return int(g_petri_nets[size_t(handle)]->places().size());
 }
 
 //------------------------------------------------------------------------------
@@ -111,13 +111,13 @@ bool petri_get_places(int const handle, CPlace_t* places)
 {
     SANITY_HANDLE(handle, false);
 
-    std::deque<Place> const& p = g_petri_nets[handle]->places();
+    std::deque<Place> const& p = g_petri_nets[size_t(handle)]->places();
     size_t i = 0;
     for (auto const& it: p)
     {
         places[i].x = it.x;
         places[i].y = it.y;
-        places[i].tokens = it.tokens;
+        places[i].tokens = int(it.tokens);
         i += 1;
     }
 
@@ -129,7 +129,7 @@ int petri_add_place(int const handle, float const x, float const y, size_t const
 {
     SANITY_HANDLE(handle, -1);
 
-    Place& p = g_petri_nets[handle]->addPlace(x, y, tokens);
+    Place& p = g_petri_nets[size_t(handle)]->addPlace(x, y, tokens);
     return int(p.id);
 }
 
@@ -138,7 +138,7 @@ int petri_add_transition(int const handle, float const x, float const y)
 {
     SANITY_HANDLE(handle, -1);
 
-    Transition& t = g_petri_nets[handle]->addTransition(x, y);
+    Transition& t = g_petri_nets[size_t(handle)]->addTransition(x, y);
     return int(t.id);
 }
 
@@ -146,7 +146,7 @@ int petri_add_transition(int const handle, float const x, float const y)
 int petri_count_transitions(int const handle)
 {
     SANITY_HANDLE(handle, -1);
-    return int(g_petri_nets[handle]->transitions().size());
+    return int(g_petri_nets[size_t(handle)]->transitions().size());
 }
 
 //------------------------------------------------------------------------------
@@ -154,7 +154,7 @@ bool petri_get_transitions(int const handle, CTransition_t* transitions)
 {
     SANITY_HANDLE(handle, false);
 
-    std::deque<Transition> const& t = g_petri_nets[handle]->transitions();
+    std::deque<Transition> const& t = g_petri_nets[size_t(handle)]->transitions();
     size_t i = 0;
     for (auto const& it: t)
     {
@@ -171,15 +171,15 @@ bool petri_remove_place(int const handle, int const id)
 {
     SANITY_HANDLE(handle, false);
 
-    if ((id < 0) || (size_t(id) >= g_petri_nets[handle]->places().size()))
+    if ((id < 0) || (size_t(id) >= g_petri_nets[size_t(handle)]->places().size()))
         return false;
 
-    std::string strid = Place::to_str(id);
-    Node* node = g_petri_nets[handle]->findNode(strid);
+    std::string strid = Place::to_str(size_t(id));
+    Node* node = g_petri_nets[size_t(handle)]->findNode(strid);
     if (node == nullptr)
         return false;
 
-    g_petri_nets[handle]->removeNode(*node);
+    g_petri_nets[size_t(handle)]->removeNode(*node);
     return true;
 }
 
@@ -188,15 +188,15 @@ bool petri_remove_transition(int const handle, int const id)
 {
     SANITY_HANDLE(handle, false);
 
-    if ((id < 0) || (size_t(id) >= g_petri_nets[handle]->transitions().size()))
+    if ((id < 0) || (size_t(id) >= g_petri_nets[size_t(handle)]->transitions().size()))
         return false;
 
-    std::string strid = Transition::to_str(id);
-    Node* node = g_petri_nets[handle]->findNode(strid);
+    std::string strid = Transition::to_str(size_t(id));
+    Node* node = g_petri_nets[size_t(handle)]->findNode(strid);
     if (node == nullptr)
         return false;
 
-    g_petri_nets[handle]->removeNode(*node);
+    g_petri_nets[size_t(handle)]->removeNode(*node);
     return true;
 }
 
@@ -205,18 +205,18 @@ int petri_add_arc(int const handle,const char* from, const char* to, float const
 {
     SANITY_HANDLE(handle, -1);
 
-    Node* node_from = g_petri_nets[handle]->findNode(from);
+    Node* node_from = g_petri_nets[size_t(handle)]->findNode(from);
     if (node_from == nullptr)
         return -1;
 
-    Node* node_to = g_petri_nets[handle]->findNode(to);
+    Node* node_to = g_petri_nets[size_t(handle)]->findNode(to);
     if (node_to == nullptr)
         return -1;
 
-    if (!g_petri_nets[handle]->addArc(*node_from, *node_to, duration))
+    if (!g_petri_nets[size_t(handle)]->addArc(*node_from, *node_to, duration))
         return -1;
 
-    return int(g_petri_nets[handle]->arcs().size() - 1u);
+    return int(g_petri_nets[size_t(handle)]->arcs().size() - 1u);
 }
 
 //------------------------------------------------------------------------------
@@ -224,15 +224,15 @@ bool petri_remove_arc(int const handle, const char* from, const char* to)
 {
     SANITY_HANDLE(handle, false);
 
-    Node* node_from = g_petri_nets[handle]->findNode(from);
+    Node* node_from = g_petri_nets[size_t(handle)]->findNode(from);
     if (node_from == nullptr)
         return false;
 
-    Node* node_to = g_petri_nets[handle]->findNode(to);
+    Node* node_to = g_petri_nets[size_t(handle)]->findNode(to);
     if (node_to == nullptr)
         return false;
 
-    return g_petri_nets[handle]->removeArc(*node_from, *node_to);
+    return g_petri_nets[size_t(handle)]->removeArc(*node_from, *node_to);
 }
 
 //------------------------------------------------------------------------------
@@ -240,10 +240,10 @@ int petri_get_tokens(int const handle, int const id)
 {
     SANITY_HANDLE(handle, -1);
 
-    auto const& places = g_petri_nets[handle]->places();
+    auto const& places = g_petri_nets[size_t(handle)]->places();
     if ((id < 0) || (id > int(places.size())))
         return -1;
-    return places[id].tokens;
+    return int(places[size_t(id)].tokens);
 }
 
 //------------------------------------------------------------------------------
@@ -251,11 +251,11 @@ bool petri_set_tokens(int const handle, int const id, size_t const tokens)
 {
     SANITY_HANDLE(handle, false);
 
-    auto& places = g_petri_nets[handle]->places();
+    auto& places = g_petri_nets[size_t(handle)]->places();
     if ((id < 0) || (id > int(places.size())))
         return -1;
 
-    places[id].tokens = tokens;
+    places[size_t(id)].tokens = tokens;
     return true;
 }
 
@@ -264,7 +264,7 @@ bool petri_save(int const handle, const char* filepath)
 {
     SANITY_HANDLE(handle, false);
 
-    return g_petri_nets[handle]->save(filepath);
+    return g_petri_nets[size_t(handle)]->save(filepath);
 }
 
 //------------------------------------------------------------------------------
@@ -272,7 +272,7 @@ bool petri_load(int const handle, const char* filepath)
 {
     SANITY_HANDLE(handle, false);
 
-    return g_petri_nets[handle]->load(filepath);
+    return g_petri_nets[size_t(handle)]->load(filepath);
 }
 
 //------------------------------------------------------------------------------
@@ -282,7 +282,7 @@ bool petri_is_event_graph(int const handle, bool* res)
     if (res == nullptr)
         return false;
 
-    *res = g_petri_nets[handle]->isEventGraph();
+    *res = g_petri_nets[size_t(handle)]->isEventGraph();
     return true;
 }
 
@@ -290,11 +290,11 @@ bool petri_is_event_graph(int const handle, bool* res)
 int petri_to_canonical(int const handle)
 {
     SANITY_HANDLE(handle, -1);
-    if (!g_petri_nets[handle]->isEventGraph())
+    if (!g_petri_nets[size_t(handle)]->isEventGraph())
         return -1;
 
     int pn = petri_create();
-    g_petri_nets[handle]->toCanonicalForm(*g_petri_nets[pn]);
+    g_petri_nets[size_t(handle)]->toCanonicalForm(*g_petri_nets[size_t(pn)]);
     return pn;
 }
 
@@ -322,10 +322,10 @@ bool petri_to_adjacency_matrices(int const handle, CSparseMatrix_t* pN, CSparseM
         return false;
     }
 
-    if (!g_petri_nets[handle]->isEventGraph())
+    if (!g_petri_nets[size_t(handle)]->isEventGraph())
         return false;
 
-    g_petri_nets[handle]->toAdjacencyMatrices(N, T);
+    g_petri_nets[size_t(handle)]->toAdjacencyMatrices(N, T);
     reference(N, pN);
     reference(T, pT);
     return true;
@@ -339,10 +339,10 @@ bool petri_to_sys_lin(int const handle, CSparseMatrix_t* pD, CSparseMatrix_t* pA
     static SparseMatrix B; static SparseMatrix C;
 
     SANITY_HANDLE(handle, false);
-    if (!g_petri_nets[handle]->isEventGraph())
+    if (!g_petri_nets[size_t(handle)]->isEventGraph())
         return false;
 
-    g_petri_nets[handle]->toSysLin(D, A, B, C);
+    g_petri_nets[size_t(handle)]->toSysLin(D, A, B, C);
     reference(D, pD);
     reference(A, pA);
     reference(B, pB);
@@ -355,10 +355,10 @@ bool petri_to_sys_lin(int const handle, CSparseMatrix_t* pD, CSparseMatrix_t* pA
 bool petri_dater_form(int const handle)
 {
     SANITY_HANDLE(handle, false);
-    if (!g_petri_nets[handle]->isEventGraph())
+    if (!g_petri_nets[size_t(handle)]->isEventGraph())
         return false;
 
-    std::cout << g_petri_nets[handle]->showDaterForm("").str() << std::endl;
+    std::cout << g_petri_nets[size_t(handle)]->showDaterForm("").str() << std::endl;
     return true;
 }
 
@@ -366,9 +366,9 @@ bool petri_dater_form(int const handle)
 bool petri_counter_form(int const handle)
 {
     SANITY_HANDLE(handle, false);
-    if (!g_petri_nets[handle]->isEventGraph())
+    if (!g_petri_nets[size_t(handle)]->isEventGraph())
         return false;
 
-    std::cout << g_petri_nets[handle]->showCounterForm("").str() << std::endl;
+    std::cout << g_petri_nets[size_t(handle)]->showCounterForm("").str() << std::endl;
     return true;
 }
